增加题库知识树、学生专业、管理员管理学生信息、登陆地点记录、选学生发布考试时，可以筛选来自不同班级的不同学生、题目支持插入音频

+++ src/main/java/cn/org/alan/exam/model/form/answer/CorrectAnswerFrom.java
@@ -22,7 +22,7 @@ public class CorrectAnswerFrom {
     // 试题ID
     @NotBlank(message = "试题Id不能为空",groups = AnswerGroup.CorrectGroup.class)
     private Integer questionId;
-    // 分数
+    // 分数（展示分，可带两位小数）
     @NotBlank(message = "分数不能为空",groups = AnswerGroup.CorrectGroup.class)
-    private Integer score;
+    private Double score;
 }
