+++ src/main/java/cn/org/alan/exam/model/vo/exam/ExamVO.java
@@ -31,14 +31,14 @@ public class ExamVO {
     private Integer examDuration;
 
     /**
-     * 及格分
+     * 及格分（展示分）
      */
-    private Integer passedScore;
+    private Double passedScore;
 
     /**
-     * 总分数
+     * 总分数（展示分）
      */
-    private Integer grossScore;
+    private Double grossScore;
 
     /**
      * 最大切屏次数
@@ -62,9 +62,9 @@ public class ExamVO {
     private Integer radioCount;
 
     /**
-     * 单选题成绩     数据库存储*100，前端正常输入和展示/100
+     * 单选题成绩（展示分）
      */
-    private Integer radioScore;
+    private Double radioScore;
 
     /**
      * 多选题数量
@@ -72,9 +72,9 @@ public class ExamVO {
     private Integer multiCount;
 
     /**
-     * 多选题成绩     数据库存储*100，前端正常输入和展示/100
+     * 多选题成绩（展示分）
      */
-    private Integer multiScore;
+    private Double multiScore;
 
     /**
      * 判断题数量
@@ -82,9 +82,9 @@ public class ExamVO {
     private Integer judgeCount;
 
     /**
-     * 判断题成绩     数据库存储*100，前端正常输入和展示/100
+     * 判断题成绩（展示分）
      */
-    private Integer judgeScore;
+    private Double judgeScore;
 
     /**
      * 简答题数量
@@ -92,9 +92,9 @@ public class ExamVO {
     private Integer saqCount;
 
     /**
-     * 简答题成绩     数据库存储*100，前端正常输入和展示/100
+     * 简答题成绩（展示分）
      */
-    private Integer saqScore;
+    private Double saqScore;
 
     /**
      * 复合题数量
@@ -102,9 +102,9 @@ public class ExamVO {
     private Integer compoundCount;
 
     /**
-     * 复合题成绩     数据库存储*100，前端正常输入和展示/100
+     * 复合题成绩（展示分）
      */
-    private Integer compoundScore;
+    private Double compoundScore;
 
     /**
      * 开始时间     YYYY-MM-DD hh:mm:ss
